--- packages/main-ui/js/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "main-ui",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "build": "npx webpack",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "caver-js": "^1.12.2",
-    "https-browserify": "^1.0.0",
-    "os-browserify": "^0.3.0",
-    "process": "^0.11.10",
-    "vm-browserify": "^1.1.2"
-  },
-  "devDependencies": {
-    "ts-loader": "^9.5.1",
-    "typescript": "^5.4.5",
-    "webpack": "^5.91.0",
-    "webpack-bundle-analyzer": "^4.10.2",
-    "webpack-cli": "^5.1.4",
-    "daisyui": "^4.12.23"
-  }
-}
