+++ packages/main-ui/package.json
@@ -8,7 +8,15 @@
   "author": "",
   "license": "ISC",
   "description": "",
+  "dependencies": {
+    "firebase": "^11.3.1"
+  },
   "devDependencies": {
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.4.5",
+    "webpack": "^5.91.0",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4",
     "daisyui": "^4.12.23"
   }
 }
