+++ packages/main-ui/src/main.rs
@@ -73,6 +73,7 @@ fn app() -> Element {
         document::Link { rel: "stylesheet", href: asset!("/public/main.css") }
         document::Link { rel: "stylesheet", href: asset!("/public/tailwind.css") }
 
+        document::Script { src: asset!("/public/dep.js") }
         document::Script { src: "https://cdn.tailwindcss.com/3.4.16" }
         document::Link {
             href: "https://cdn.jsdelivr.net/npm/daisyui@4.12.23/dist/full.min.css",
