+++ packages/main-ui/webpack.config.js
@@ -1,0 +1,33 @@
+const webpack = require("webpack");
+const path = require("path");
+
+module.exports = {
+  output: {
+    globalObject: "this",
+    filename: "dep.js",
+    path: path.resolve(__dirname, "public"),
+  },
+  entry: "./js/index.ts",
+  devtool: "inline-source-map",
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: "ts-loader",
+        exclude: [/node_modules/],
+      },
+    ],
+  },
+  experiments: {
+    asyncWebAssembly: true,
+    syncWebAssembly: true,
+  },
+  resolve: {
+    extensions: [".tsx", ".ts", ".js"],
+  },
+  plugins: [
+    new webpack.ProvidePlugin({
+      Buffer: ["buffer", "Buffer"],
+    }),
+  ],
+};
